Rename typo in file name
